requirements.txt
numpy==1.24.3
pandas==1.5.3
pillow==9.4.0
requests==2.29.0
scipy
tensorboard==2.12.1
torch
torchaudio
torchvision
tqdm==4.65.0
transformers==4.29.2
werkzeug==2.2.3
addict==2.4.0
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
configargparse==1.5.3
contourpy==1.1.0
cycler==0.11.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.6.1
executing==1.2.0
fastjsonschema==2.17.1
flask==2.2.5
fonttools==4.40.0
gitdb==4.0.10
gitpython==3.1.31
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.12.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lightning-utilities==0.9.0
markupsafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.6
objaverse==0.0.7
open3d
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
platformdirs==3.7.0
plotly==5.15.0
plyfile==0.9
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocoevalcap==1.2
pycocotools==2.0.6
pygments==2.15.1
pyparsing==3.1.0
pyquaternion==0.9.9
pyrsistent==0.19.3
pytorch-lightning==2.0.4
pyzmq==25.1.0
scikit-learn==1.2.2
sentry-sdk==1.25.1
setproctitle==1.3.2
smmap==5.0.0
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
torchmetrics==0.11.4
tornado==6.3.2
traitlets==5.9.0
trimesh==3.22.1
wandb==0.15.4
widgetsnbextension==4.0.7
zipp==3.15.0
omegaconf==2.3.0
hydra-core==1.3.2
